pyproject.toml
[project]
name = "vpn-manager"
version = "0.1.0"
description = "Telegram bot for managing VPN subscriptions via 3x-ui panels"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "VPN Manager Team"}
]
keywords = ["vpn", "telegram", "bot", "3x-ui", "xray"]

dependencies = [
    "aiogram>=3.4.0",
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.0",
    "aiohttp>=3.9.0",
    "pydantic-settings>=2.1.0",
    "cryptography>=41.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
